django/requirements.txt
Django
djangorestframework
django-filter
psycopg2  # PostgreSQL driver
